% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directional_outlyingness.R
\name{dir_out}
\alias{dir_out}
\title{Compute the directional outlyingness}
\usage{
dir_out(
  data,
  dirout_matrix = FALSE,
  data_depth = c("MhD", "RP", "SD", "HS"),
  return_distance = F
)
}
\arguments{
\item{data}{A matrix for univariate functional data or a 3d array for multivariate functional data.}

\item{dirout_matrix}{If TRUE, returns the directional outlyingness matrix.}

\item{data_depth}{The method for computing the depth- mahalanobis, RP, SD, or Half-space depth.}

\item{return_distance}{If TRUE, returns the the distance values of the matrix whose rows are the mean and variation of directional outlyiness.}
}
\description{
Compute the the directional outlyingness of a univariate or multivariate functional data.
}
