% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directional_quantile.R
\name{directional_quantile}
\alias{directional_quantile}
\title{Title}
\usage{
directional_quantile(dt, quantiles = c(0.025, 0.975))
}
\arguments{
\item{dt}{A matrix or dataframe of size \eqn{n} observations/curves by \eqn{p} domain/evaluation
points.}

\item{quantiles}{A numeric vector of length 2 specifying the probabilities of the lower and upper quantiles.
Defaults to \code{c(0.025, 0.975)}}
}
\value{
A numeric vector containing the negative of the directional quantile.
}
\description{
Title
}
