% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extreme_rank_length.R
\name{extreme_rank_length}
\alias{extreme_rank_length}
\title{Extreme Rank Length Depth}
\usage{
extreme_rank_length(
  dt,
  type = c("two_sided", "one_sided_left", "one_sided_right")
)
}
\arguments{
\item{dt}{A matrix of n curves and p observations points}

\item{type}{}
}
\value{
A numeric vector containing the depth of each curve
}
\description{
Extreme Rank Length Depth
}
\examples{

}
