% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hardin_rocke_f_factors.R
\name{hardin_factor_asymptotic}
\alias{hardin_factor_asymptotic}
\title{Compute the factors for Hardin's F approximation cutoff asymptotically}
\usage{
hardin_factor_asymptotic(n, dimension)
}
\arguments{
\item{n}{A numeric value indicating the number of observations of the data}

\item{dimension}{A numeric value indicating the number of columns or variables.}
}
\value{

}
\description{
Compute the factors for Hardin's F approximation cutoff asymptotically
}
