% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muod.R
\name{muod}
\alias{muod}
\title{Massive Unspervised Outlier Detection (MUOD)}
\usage{
muod(data, cut_method = c("boxplot", "tangent"))
}
\arguments{
\item{data}{a matrix or dataframe of size \code{n} observation by \code{p} domain points.}

\item{cut_method}{a character value indicating method to use for finding indices cutoff.
Must be either \code{"boxplot"} or \code{"tangent"}.}
}
\value{
Returns a list containing the following
  \item{\code{outliers}}{a vector containing the indices of outliers identified.}
  \item{\code{indices}}{a dataframe containing the shape, magnitude and amplitude indices}
}
\description{
Massive Unspervised Outlier Detection (MUOD)
}
