% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_variation_depth.R
\name{total_variation_depth}
\alias{total_variation_depth}
\title{Compute total variation depth}
\usage{
total_variation_depth(data, n_curves = nrow(data), n_points = ncol(data))
}
\arguments{
\item{n_points}{}
}
\value{

}
\description{
Compute total variation depth
}
