% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvdmss.R
\name{tvd_mss}
\alias{tvd_mss}
\title{Find shape and magnitude outliers  using the Total Variation Depth
 and Modified Shape Similarity Index}
\usage{
tvd_mss(data, n_curves = nrow(data), n_points = ncol(data), emp_factor = 1.5)
}
\arguments{
\item{data}{A matrix of size \eqn{n} observations by \eqn{p} domain
points.}

\item{n_curves}{The number of observations or curves. Set to \code{nrow(data)}
by default.}

\item{n_points}{The number of domain/evaluation points. Set to \code{ncol(data)}
by default.}

\item{emp_factor}{The empirical factor of the boxplot used on the modified shape
similarity index.}
}
\value{

}
\description{
Find shape and magnitude outliers  using the Total Variation Depth
 and Modified Shape Similarity Index
}
