% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{world_population}
\alias{world_population}
\title{World Population Data by Countries}
\format{
A matrix of size 105 rows by 61 columns.
}
\source{
Data originally available in the \code{depth.fd} package.
}
\usage{
world_population
}
\description{
This is the world population data, revision 2010, by countries used in the paper
Nagy et al. (2016) \doi{10.1080/10618600.2017.1336445} and Dai et al. (2020)
\doi{10.1016/j.csda.2020.106960}. It contains population (both sexes) of countries
as of July 1 in the years 1950 - 2010. The data have been pre-processed as described
in Nagy et al. (2016) and hence contains only the 105 countries with population
in the range of one million and fifteen million on July 1, 1980.
}
\details{
Data included for illustration and testing purposes.
}
\examples{
data(world_population)
str(world_population)

}
\keyword{datasets}
