% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msplot.R
\name{ms_plot}
\alias{ms_plot}
\title{Magniude-Shape Plot (MS-Plot) based on the directional outlyingness for functional data.}
\usage{
ms_plot(
  data,
  plot = F,
  plot_type = c("scatter", "parallel"),
  return_outliers = T,
  data_depth = c("random_projections", "mahalanobis", "simplicial", "half_space"),
  normal_col = "blue",
  outlier_col = "red",
  col = "grey"
)
}
\arguments{
\item{data}{A matrix for univariate functional data (of size n observations
by p domain points) or a 3-dimensional array for multivariate functional
data (of size n observations by p domain points by d dimension).}

\item{plot}{A logical scalar. If TRUE, makes the MS-Plot}

\item{plot_type}{The type of plot to produce if \code{data} is a multivariate functional data.
Can be one of "scatter" or "parallel". Default is "Scatter". For univariate functional data, the
scatter plot is always produced.}

\item{return_outliers}{A logical scalar. IF TRUE, returns the indices of the outliers. Furthermore, if
\code{plot} is TRUE, differentiates the outlying points of the MS-Plot from those of the non-outlying
points.}

\item{data_depth}{The depth used in the computation of the directional outlyingness if \code{data} is a
multivariate functional data. Can be one of "mahalanobis", "random_projections", "simplicial", or "half_space" depth.
Default is "random_projections". For univariate functional data, the projection depth is always used.}

\item{normal_col}{The color of non-outlying points in the MS-Plot. Must be a character vector
#' containing one name of a standard color in R or the Hex code of a color.}

\item{outlier_col}{The color of outlying points in the MS-Plot. Must be a character vector
#' containing one name of a standard color in R or the Hex code of a color.}

\item{col}{The color for plotting points in the MS-Plot when \code{return_outliers} is FALSE and
\code{plot} is TRUE.}
}
\value{

}
\description{
This functiona creates the MS-Plot, an outlier detection tool for univariate and
multivariate functional data proposed in Wenlin and Genton (2019). Indices of observations
flagged as outliers are returned, together with the corresponding plot which depends on the
 dimension of the data (see details).
}
\details{
This function implements the Magnitude-Shape Plot of Dai and Genton (2018). MS-Plot is a plot of the variation
of directional outlyingness (VO) against the mean directional outlyingness (MO). MO and VO are computed based on
the directional outlyingness defined in Dai and Genton (2019). For univariate functional data,
the projection depth based on Zuo (2003) is always used (as suggested by Dai and Genton (2019)) for
computing the directional outlyingess while for multivariate functional data, any of "mahalanobis", "random_projections",
"simplicial", or "half_space" depths can be used.

For univariate functional data, a 2-d scatter plot is always produced using ggplot2. This plot is one of the objects
contained in the list returned by the function. For bivariate functional data, a 3-d scatter plot (made with the scatter3d package)
or a parallel plot (made with ggplot2) is produced. For multivariate functional data (with dimensions >= 2), a 2-d scatter plot
(of VO against the norm of MO) or a parallel plot (both returned as a ggplot2 object) is produced.
}
\examples{

}
\references{
Dai, W., and Genton, M. G. (2018). Multivariate functional data visualization and outlier detection. \emph{Journal of Computational and Graphical Statistics}, 27(4), 923-934.

Dai, W., and Genton, M. G. (2019). Directional outlyingness for multivariate functional data. \emph{Computational Statistics & Data Analysis}, 131, 50-65.
}
\seealso{

}
\author{
Version created by Oluwasegun Taiwo Ojo based on the original code written by Wenlin Dai and Marc G. Genton.
}
