% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_projection_depth.R
\name{projection_depth}
\alias{projection_depth}
\title{Random Projection for Multivariate Data}
\usage{
projection_depth(dt, n_projections = 500L, seed = NULL, delta = 1e-15)
}
\arguments{
\item{dt}{A matrix or dataframe of size n observations by d dimension).}

\item{n_projections}{An integer, indicating the number of directions for random projections. By default,
500 random directions for projection are generated from a scaled uniform distribution.}

\item{seed}{An integer to set a seed when generating the random directions. Useful for reproducibility.}

\item{delta}{Accuracy. The left limit of the empirical distribution function.}
}
\value{
A vector contaning the depth values \code{dt} with respect to itself.
}
\description{
Helper function to compute the random projection depth of a multivariate data with respect to itself.
}
\author{
Version adapted from original code contained in \code{fda.usc}.
}
