% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvdmss.R
\name{tvd_mss}
\alias{tvd_mss}
\title{Find shape and magnitude outliers  using the Total Variation Depth
 and Modified Shape Similarity Index}
\usage{
tvd_mss(data, emp_factor = 1.5)
}
\arguments{
\item{data}{A matrix or dataframe of size \eqn{n} observations/curves by \eqn{p} domain/evaluation
points.}

\item{emp_factor}{The empirical factor of the boxplot used on the modified shape
  similarity index.

This method uses a combination of total variation depth (TVD) and modified shape similarity (MSS) index
 defined in Huang and Sun (2019) to find magnitude and shape outliers. The MSS of the observations are
 first computed and a classical boxplot applied on the indices. Outliers detected by the boxplot are
 flagged as shape outliers. The shape outliers are then removed from the data and then TVD of the
 remaning observations are used in a functional boxplot to detect magnitude outliers.}
}
\value{
Returns a list contaning the following
\describe{
  \item{\code{outliers}}{the indices of the (shape and magnitude) outliers}
  \item{\code{shape_outliers}}{the indices of the shape outliers}
  \item{\code{magnitude_outliers}}{the indices of the magnitude outliers}
  \item{\code{tvd}}{the total variation depths of the observations of \code{data}}
  \item{\code{mss}}{the modified shape similarity index of the observations of \code{data}}
  }
}
\description{
Find shape and magnitude outliers  using the Total Variation Depth
 and Modified Shape Similarity Index
}
\examples{
data(sim_data1)
res <- tvd_mss(sim_data1$data)
}
\references{
Huang, H., & Sun, Y. (2019). A decomposition of total variation depth for
understanding functional outliers. \emph{Technometrics}, 61(4), 445-458.
}
\seealso{
\code{\link{msplot}} for outlier detection using msplot.
}
\author{
Oluwasegun Ojo
}
